image-service/requirements.txt
fastapi
uvicorn
pillow
aiohttp
